reglement differe

+++ src/main/java/com/kobe/warehouse/service/reglement/differe/dto/CustomerReglementDiffereDTO.java
@@ -3,7 +3,7 @@ package com.kobe.warehouse.service.reglement.differe.dto;
 
 
 
-public record CustomerReglementDiffereDTO(Long id,  String firstName, String lastName, int paidAmount) {
+public record CustomerReglementDiffereDTO(Long id,  String firstName, String lastName, long paidAmount) {
 
 
 }
