+++ src/main/java/com/kobe/warehouse/service/reglement/differe/dto/ReglementDiffereWrapperDTO.java
@@ -3,20 +3,21 @@ package com.kobe.warehouse.service.reglement.differe.dto;
 
 import com.kobe.warehouse.service.utils.NumberUtil;
 
-import java.math.BigDecimal;
 import java.util.List;
 
-public record ReglementDiffereWrapperDTO(Long id,  String firstName, String lastName, int paidAmount, Long solde,
-                                         List<ReglementDiffereDTO>items) {
+public record ReglementDiffereWrapperDTO(Long id, String firstName, String lastName, long paidAmount, Long solde,
+                                         List<ReglementDiffereDTO> items) {
 
     public String customerfullName() {
         return String.format("%s %s", firstName, lastName);
     }
+
     public String formattedPaidAmount() {
         return NumberUtil.formatToString(paidAmount);
     }
+
     public String formattedSolde() {
-        if(solde == null) {
+        if (solde == null) {
             return "";
         }
         return NumberUtil.formatToString(solde);
